fix(core): resolve lint errors in principles-core runtime-v2
- Replace non-null assertions with safe access patterns
- Move helper functions before usage to fix use-before-define errors
- Refactor rankCandidates/runTournament to use options objects (max-params)
- Use object destructuring in snapshot-contract validation
- Update all call sites for new function signatures
