+++ packages/principles-core/src/runtime-v2/nocturnal/index.ts
@@ -1,0 +1,57 @@
+export type {
+  ArtificerTargetRuleScore,
+  ArtificerTargetRuleResolution,
+  TrinityArtificerContext,
+  TrinityDreamerCandidate,
+  TrinityDreamerOutput,
+  PhilosopherRiskAssessment,
+  Philosopher6DScores,
+  PhilosopherJudgment,
+  PhilosopherOutput,
+  TournamentTraceEntry,
+  RejectedAnalysis,
+  ChosenJustification,
+  ContrastiveAnalysis,
+  TrinityTelemetry,
+  TrinityStageFailure,
+  TrinityResult,
+  TrinityDraftArtifact,
+} from './nocturnal-trinity-types.js';
+
+// Candidate scoring (migrated from openclaw-plugin)
+export type {
+  ThresholdValues,
+  CandidateScores,
+  ScoredCandidate,
+  TournamentResult,
+  TournamentTraceEntry as CandidateTournamentTraceEntry,
+  ScoringWeights,
+  DiversityValidationResult,
+  RankCandidatesOptions,
+  RunTournamentOptions,
+} from './candidate-scoring.js';
+
+export {
+  DEFAULT_SCORING_WEIGHTS,
+  scoreCandidate,
+  checkThresholds,
+  validateCandidateDiversity,
+  rankCandidates,
+  runTournament,
+} from './candidate-scoring.js';
+
+// Snapshot contract (migrated from openclaw-plugin)
+export type {
+  NocturnalAssistantTurn,
+  NocturnalUserTurn,
+  NocturnalToolCall,
+  NocturnalPainEvent,
+  NocturnalGateBlock,
+  NocturnalUserCorrection,
+  NocturnalSessionSnapshot,
+  NocturnalSnapshotContractResult,
+} from './snapshot-contract.js';
+
+export {
+  validateNocturnalSnapshotIngress,
+} from './snapshot-contract.js';
