+++ packages/principles-core/src/runtime-v2/types/principle-dependency.ts
@@ -1,0 +1,6 @@
+export interface PrincipleDependency {
+  principleId: string;
+  dependsOn: string[];
+  conflictedWith: string[];
+  supersedes: string[];
+}
