+++ packages/principles-core/src/runtime-v2/types/principle-value-metrics.ts
@@ -1,0 +1,12 @@
+export interface PrincipleValueMetrics {
+  principleId: string;
+  painPreventedCount: number;
+  avgPainSeverityPrevented: number;
+  lastPainPreventedAt?: string;
+  totalOpportunities: number;
+  adheredCount: number;
+  violatedCount: number;
+  implementationCost: number;
+  benefitScore: number;
+  calculatedAt: string;
+}
