+++ packages/principles-core/src/runtime-v2/types/runtime-summary-types.ts
@@ -1,0 +1,49 @@
+/**
+ * Runtime truth represents the current state of the system.
+ * Used for control decisions, Phase 3 eligibility, and real-time operations.
+ * Sources: queue state, active session registry
+ */
+export interface RuntimeTruth {
+  queueState: {
+    total: number;
+    pending: number;
+    inProgress: number;
+    completed: number;
+    lastUpdated: string;
+  };
+  activeSessions: string[];
+}
+
+/**
+ * Analytics truth represents historical data and aggregated metrics.
+ * Used for insights, trends, and supporting evidence (where explicitly allowed).
+ * NOT used for control decisions or Phase 3 eligibility.
+ * Sources: trajectory.db, daily-stats.json, control-ui DB
+ */
+export interface AnalyticsTruth {
+  trajectoryData: {
+    totalTasks: number;
+    successRate: number;
+    timeoutRate: number;
+    lastUpdated: string;
+  };
+  dailyStats: {
+    toolCalls: number;
+    painSignals: number;
+    evolutionTasks: number;
+    lastUpdated: string;
+  };
+  trends: {
+    sevenDay: TrendMetrics;
+    thirtyDay: TrendMetrics;
+  };
+}
+
+/**
+ * Trend metrics for analytics aggregation.
+ */
+export interface TrendMetrics {
+  successRateChange: number;
+  toolCallVolumeChange: number;
+  painSignalRateChange: number;
+}
